pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tarxemo-django-authz"
version = "0.1.0"
authors = [
  { name="TarXemo" },
]
description = "A comprehensive, reusable Django authorization framework supporting RBAC and PBAC"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Django>=3.2",
    "graphene-django>=3.0",
    "tarxemo-django-graphene-utils>=0.1.2",
]

[project.optional-dependencies]
drf = ["djangorestframework>=3.12"]

[project.urls]
"Homepage" = "https://github.com/tarxemo/tarxemo-django-authz"
"Bug Tracker" = "https://github.com/tarxemo/tarxemo-django-authz/issues"
